Add curl scripts for the games
Added curl scripts and a folder for the JS for games.
