+++ assets/curl-scripts/games/index-games.sh
@@ -1,0 +1,5 @@
+curl "https://tic-tac-toe-wdi.herokuapp.com/games" \
+  --include \
+  --request GET \
+  --header "Content-Type: application/json" \
+  --header "Authorization: Token token=${TOKEN}" \
