+++ assets/scripts/app.js
@@ -3,9 +3,11 @@
 // use require with a reference to bundle the file and use it in this file
 // const example = require('./example')
 const authEvents = require('./auth/events.js')
+const gameEvents = require('./game/events.js')
 // use require without a reference to ensure a file is bundled
 // require('./example')
 
 $(() => {
-  authEvents.addHandlers()
+  authEvents.addHandlers(),
+  gameEvents.addHandlers()
 })
