+++ assets/scripts/games/ui.js
@@ -1,0 +1,23 @@
+'use strict'
+
+const store = require('../store')
+
+const selectSuccess = function (data) {
+  $('#message').text('Signed in successfully')
+  $('#message').removeClass()
+  $('#message').addClass('success')
+  console.log('signInSuccess ran. Data is :', data)
+  store.user = data.user
+}
+
+const selectFailure = function (error) {
+  $('#message').text('Error on sign in')
+  $('#message').removeClass()
+  $('#message').addClass('failure')
+  console.error('signInFailure ran. Error is :', error)
+}
+
+module.exports = {
+  selectSuccess,
+  selectFailure
+}
