+++ assets/styles/index.scss
@@ -1,7 +1,7 @@
 @import '~bootstrap/scss/bootstrap';
 .box {
   border: 2px solid black;
-  height: 100px;
+  height: 200px;
   display: flex;
   align-items: center;
 }
