+++ index.html
@@ -54,7 +54,7 @@
             </fieldset>
           </form>
           <div class="container">
-            <div class="gameboard">
+            <div id="gameboard">
               <div class="row">
                 <div class="col-4 box"><span class="border-top-0 border-left-0">Box 0</span></div>
                 <div class="col-4 box"><span class="border-top-0">Box 1</span></div>
